+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.9.16
 
+* added `terminating` property to `TerminableThread`
+* added `trace_function`